Move the IRequestExecutor to APIContext
